move components to base and section directories
