+++ gatsby-theme-landing-page/src/components/base/base.module.css
