+++ gatsby-theme-landing-page/src/components/base/head.js
