+++ gatsby-theme-landing-page/src/components/base/layout.js
