+++ gatsby-theme-landing-page/src/components/index.js
@@ -1,5 +1,5 @@
-export { default as Hero } from "./hero";
-export { default as Features } from "./features";
-export { default as Copy } from "./copy";
-export { default as CallToAction } from "./call-to-action";
-export { default as Benefits } from "./benefits";
+export { default as Hero } from "./sections/hero";
+export { default as Features } from "./sections/features";
+export { default as Copy } from "./sections/copy";
+export { default as CallToAction } from "./sections/call-to-action";
+export { default as Benefits } from "./sections/benefits";
