+++ gatsby-theme-landing-page/src/components/sections/copy.js
@@ -1,6 +1,6 @@
 import React from "react";
 import * as styles from "./copy.module.css";
-import MarkdownText from "./markdown-text";
+import MarkdownText from "../base/markdown-text";
 import { GatsbyImage, getImage } from "gatsby-plugin-image";
 
 export default function Copy({ heading, secondaryHeading, content }) {
