+++ gatsby-theme-landing-page/src/pages/{ContentfulLandingPage.slug}.js
@@ -2,9 +2,9 @@ import * as React from "react";
 import { graphql } from "gatsby";
 import * as Components from "../components";
 import * as styles from "../components/base.module.css";
-import Head from "../components/head";
-import Layout from "../components/layout";
-import DevDebug from "../components/dev-debug";
+import Head from "../components/base/head";
+import Layout from "../components/base/layout";
+import DevDebug from "../components/base/dev-debug";
 
 export default function LandingPage(props) {
   const { sections } = props.data.page;
